Feature #68 Copy files from /resources/dockerfile to Storage Root when the app starts
