+++ src/main/java/com/online/compiler/runnerapi/core/ApplicationReadyListener.java
@@ -1,0 +1,7 @@
+package com.online.compiler.runnerapi.core;
+
+import org.springframework.stereotype.Component;
+
+@Component
+public class ApplicationReadyListener {
+}
