+++ src/main/java/com/online/compiler/runnerapi/core/JavaVersionsManager.java
@@ -1,0 +1,7 @@
+package com.online.compiler.runnerapi.core;
+
+public class JDKVersionsManager {
+
+    
+
+}
