+++ src/main/java/com/online/compiler/runnerapi/features/runner/service/FileWriter.java
@@ -12,7 +12,7 @@ import java.nio.file.Path;
 import java.util.UUID;
 import java.util.concurrent.CompletableFuture;
 
-import static com.online.compiler.runnerapi.common.Constants.*;
+import static com.online.compiler.runnerapi.core.Constants.*;
 import static java.nio.file.StandardOpenOption.CREATE;
 import static java.nio.file.StandardOpenOption.WRITE;
